SSCS-8049 Package move and rename for dynamic benefit type about to start handler.   PR review feedback changes.   Removal of StringUtils and replacement with commons version
