+++ build.gradle
@@ -275,7 +275,7 @@ dependencies {
     compile group: 'uk.gov.hmcts.reform', name: 'logging-appinsights', version: versions.reformLogging
     compile group: 'uk.gov.hmcts.reform', name: 'doc-assembly-client', version: '1.0.6'
     compile group: 'uk.gov.hmcts.reform', name: 'document-management-client', version: '7.0.0'
-    compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '3.4.57-RC-SSCS-7935'
+    compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '3.4.57'
     compile group: 'uk.gov.hmcts.reform', name: 'sscs-pdf-email-common', version: '1.4.9'
 
     testCompile group: 'io.rest-assured', name: 'rest-assured'
