+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/abouttostart/BenefitTypeDynamicListItems.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit.dwpuploadresponse;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.abouttostart;
 
 import java.util.Arrays;
 import java.util.Optional;
